+++ starkinfra/issuingcard/__issuingcard.py
@@ -84,7 +84,7 @@ def query(status=None, types=None, holder_ids=None, after=None, before=None, tag
     """# Retrieve IssuingCards
     Receive a generator of IssuingCards objects previously created in the Stark Infra API
     ## Parameters (optional):
-    - status [string, default None]: filter for status of retrieved objects. ex: "paid" or "registered"
+    - status [string, default None]: filter for status of retrieved objects. ex: "active", "blocked", "expired" or "canceled"
     - types [list of strings, default None]: card type. ex: ["virtual"]
     - holder_ids [list of strings]: card holder IDs. ex: ["5656565656565656", "4545454545454545"]
     - after [datetime.date or string, default None] date filter for objects created only after specified date. ex: datetime.date(2020, 3, 10)
@@ -117,7 +117,7 @@ def page(status=None, types=None, holder_ids=None, after=None, before=None, tags
     """# Retrieve paged IssuingCards
     Receive a list of IssuingCards objects previously created in the Stark Infra API and the cursor to the next page.
     ## Parameters (optional):
-    - status [string, default None]: filter for status of retrieved objects. ex: "paid" or "registered"
+    - status [string, default None]: filter for status of retrieved objects. ex: "active", "blocked", "expired" or "canceled"
     - types [list of strings, default None]: card type. ex: ["virtual"]
     - holder_ids [list of strings, default None]: card holder IDs. ex: ["5656565656565656", "4545454545454545"]
     - after [datetime.date or string, default None] date filter for objects created only after specified date. ex: datetime.date(2020, 3, 10)
@@ -154,7 +154,7 @@ def get(id, expand=None, user=None):
     ## Parameters (required):
     - id [string]: object unique id. ex: "5656565656565656"
     ## Parameters (optional):
-    - expand [list of strings, default None]: fields to expand information. ex: ["rules"]
+    - expand [list of strings, default None]: fields to expand information. ex: ["rules", "security_code", "number", "expiration"]
     - user [Organization/Project object, default None]: Organization or Project object. Not necessary if starkinfra.user was set before function call
     ## Return:
     - IssuingCards object with updated attributes
@@ -168,7 +168,7 @@ def update(id, status=None, display_name=None, rules=None, tags=None, user=None)
     ## Parameters (required):
     - id [string]: IssuingCard id. ex: '5656565656565656'
     ## Parameters (optional):
-    - status [string]: You may block the IssuingCard by passing 'blocked' in the status
+    - status [string]: You may block the IssuingCard by passing 'blocked' or activate by passing 'active' in the status
     - display_name [string, default None]: card displayed name
     - rules [list of dictionaries, default None]: list of dictionaries with "amount": int, "currencyCode": string, "id": string, "interval": string, "name": string pairs.
     - tags [list of strings]: list of strings for tagging